+++ tests/test_graph.py
@@ -152,3 +152,23 @@ class TestGraph(unittest.TestCase):
         graph.add_directed_edge("A", "Y")
         graph.add_directed_edge("B", "Z")
         self.assertEqual({"A", "B", "C", "D"}, graph.ancestors_inclusive({"A", "B"}))
+
+        graph = NxMixedGraph()
+        graph.add_directed_edge("X", "Z")
+        graph.add_directed_edge("Z", "Y")
+        graph.add_undirected_edge("X", "Y")
+        self.assertEqual({"X", "Y", "Z"}, graph.ancestors_inclusive({"Y"}))
+        self.assertEqual({"X", "Z"}, graph.ancestors_inclusive({"Z"}))
+        self.assertEqual({"X"}, graph.ancestors_inclusive({"X"}))
+
+    def test_get_c_components(self):
+        """Test that get_c_components works correctly."""
+        g1 = NxMixedGraph().from_edges(directed=[("X", "Y"), ("Z", "X"), ("Z", "Y")])
+        c1 = [frozenset("X"), frozenset("Y"), frozenset("Z")]
+        g2 = NxMixedGraph().from_edges(directed=[("X", "Y")], undirected=[("X", "Y")])
+        c2 = [frozenset(["X", "Y"])]
+        g3 = NxMixedGraph().from_edges(directed=[("X", "M"), ("M", "Y")], undirected=[("X", "Y")])
+        c3 = [frozenset(["X", "Y"]), frozenset("M")]
+        for g, c in [(g1, c1), (g2, c2), (g3, c3)]:
+            self.assertIsInstance(g, NxMixedGraph)
+            self.assertEqual(c, g.get_c_components())
